refactor: standardize truth handling in evaluation and leakage checks
- Updated the evaluation and leakage check functions to use `test.require_truth()` for accessing particle-level truth, ensuring consistency across the codebase.
- Modified the plotting function to utilize `test.require_truth()` for improved clarity and correctness in particle-level comparisons.
- Enhanced the `Populations` class with methods to manage truth data, including a sentinel for absent truth, improving robustness in handling datasets without truth information.
- Adjusted type hints and documentation to reflect these changes, enhancing type safety and clarity in the data handling processes.

+++ tests/test_completion_logging.py
@@ -150,7 +150,7 @@ def test_ibu_records_metric_and_weight_artifact_completion(
         return ibu.IBUResult(
             metrics={"detector_mass": metric_record},
             variable_names=("mass", "momentum"),
-            weights=np.array([[1.0, 2.0], [3.0, 4.0]], dtype=np.double),
+            weights=np.array([[1.0, 2.0], [3.0, 4.0]], dtype=np.single),
             outcomes=(
                 ibu.VariableOutcome(
                     variable_name="mass",
